% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedPower_BC.R
\name{MedPower_BC}
\alias{MedPower_BC}
\title{Sample size or power calculation for mediation analysis (variable type BC)}
\usage{
MedPower_BC(
  alpha = 0.05,
  pA = 0.5,
  pM = 0.5,
  sigmaY = 1,
  theta0,
  theta1,
  beta1,
  theta2,
  theta3 = 0,
  n = NULL,
  power = NULL
)
}
\arguments{
\item{alpha}{Significance level. Default is 0.05.}

\item{pA}{Proportion of treated or exposed individuals. Default is 0.5.}

\item{pM}{Marginal probability of the binary mediator being 1. Default is 0.5.}

\item{sigmaY}{Standard deviation of the continuous outcome. Default is 1.}

\item{theta0}{Intercept in the outcome model.}

\item{theta1}{Effect of exposure on outcome.}

\item{beta1}{Effect of exposure on binary mediator (log-odds scale).}

\item{theta2}{Effect of mediator on outcome.}

\item{theta3}{Interaction effect between exposure and mediator on outcome.}

\item{n}{Sample size. If specified, the function returns the corresponding power.}

\item{power}{Target power. If specified, the function returns the required sample size.}
}
\value{
A list containing either:
\describe{
\item{required_n}{Sample size required to achieve the target power (if \code{power} is specified).}
\item{achieved_power}{Power achieved with the given sample size (if \code{n} is specified).}
}
}
\description{
This function computes either the statistical power or the required sample size
for detecting a natural indirect effect in a mediation model where the mediator
is binary and the outcome is continuous (variable type BC).
}
\examples{
MedPower_BC(
  theta0 = -4,
  theta1 = log(2),
  beta1 = 0.39,
  theta2 = log(3) * 4 / 5,
  theta3 = log(3) * 1 / 5,
  power = 0.8
)
}
